auth_app/requirements.txt
django
djangorestframework
django-oauth-toolkit